frontend: creation layout user for navbar user different  navbar admin, update of folder structure. And add back home  button
