+++ app/_components/NavBarAdmin.tsx
@@ -1,0 +1,8 @@
+export default function NavBarAdmin() {
+  return (
+    <div className="">
+      <p>NavbarAdmin</p>
+    </div>
+  );
+}
+
