+++ app/admin/page.tsx
@@ -1,11 +1,10 @@
-import NavBar from "./_components/NavBar";
-
 export default function DashboardAdmin() {
   return (
-    <>
-      <h2>Dashboard</h2>
-      <NavBar />
-    </>
+    <div>
+      <main>
+        <p className="mt-40">Dashboard</p>
+      </main>
+    </div>
   );
 }
