+++ app/layout.tsx
@@ -1,7 +1,7 @@
+// app/layout.tsx
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 
-import Navbar from "./_components/Navbar";
 import "./globals.css";
 
 const inter = Inter({ subsets: ["latin"] });
@@ -19,10 +19,8 @@ export default function RootLayout({
   return (
     <html lang="en" className="bg-[#fdfdfd]">
       <body className={inter.className}>
-        <header className="fixed w-screen top-0 z-50 bg-none bg-transparent  ">
-          <Navbar />
-        </header>{" "}
-        {children}
+        <div className="fixed w-screen top-0 z-50 bg-none bg-transparent"></div>
+        <main> {children}</main>
       </body>
     </html>
   );
