--- app/tarifs/page.tsx
@@ -1,8 +1,0 @@
-export default function Rates() {
-  return (
-    <div className="mt-40 ml-20">
-      <h2>Les tarifs</h2>
-    </div>
-  );
-}
-
