+++ app/user/(chambres)/chambre-1/page.tsx
