+++ app/user/(chambres)/chambre-2/page.tsx
