+++ app/user/(chambres)/chambre-3/page.tsx
