+++ app/user/jardin/page.tsx
@@ -1,0 +1,8 @@
+export default function Garden() {
+  return (
+    <>
+      <h2 className="mt-40">Jardin</h2>
+    </>
+  );
+}
+
